requirements.txt
fastapi
uvicorn
websockets
openai-whisper
numpy
pyaudio
selenium